Update heatmaps.csv
Aniq added additional columns 
The file now has columns labelled by numbers. These numbers represent a value in a slider. When users move a slider, the heatmap should ideally swap data from one column to the respective column. This is for the “A slider to change the mitigation cost for the heatmap”